requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
backports.zoneinfo==0.2.1
certifi==2022.12.7
Django==4.1.5
django-crontab==0.7.1
django-redis==5.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
mysql-client==0.0.1
mysqlclient==2.1.1
psycopg2-binary==2.9.5
PyJWT==2.6.0
PyMySQL==1.0.2
pytz==2022.7
redis==4.4.1
sqlparse==0.4.3
